float all the things

+++ parser/mesos/slave.go
@@ -117,8 +117,8 @@ func (mp SlaveParser) getSystemPoint(stats SlaveStats) store.Point {
 			"load_15min":      stats.System_load15min,
 			"load_1min":       stats.System_load1min,
 			"load_5min":       stats.System_load5min,
-			"mem_free_bytes":  stats.System_memFreeBytes,
-			"mem_total_bytes": stats.System_memTotalBytes,
+			"mem_free_bytes":  int(stats.System_memFreeBytes),
+			"mem_total_bytes": int(stats.System_memTotalBytes),
 		},
 		Time: stats.Time,
 	}
@@ -162,52 +162,52 @@ func (mp SlaveParser) getGlobalPoint(stats SlaveStats) store.Point {
 }
 
 type SlaveStats struct {
-	Containerizer_mesos_containerDestroyErrors int     `json:"containerizer/mesos/container_destroy_errors"`
-	Slave_containerLaunchErrors                int     `json:"slave/container_launch_errors"`
+	Containerizer_mesos_containerDestroyErrors float64 `json:"containerizer/mesos/container_destroy_errors"`
+	Slave_containerLaunchErrors                float64 `json:"slave/container_launch_errors"`
 	Slave_cpusPercent                          float64 `json:"slave/cpus_percent"`
 	Slave_cpusRevocablePercent                 float64 `json:"slave/cpus_revocable_percent"`
-	Slave_cpusRevocableTotal                   int     `json:"slave/cpus_revocable_total"`
-	Slave_cpusRevocableUsed                    int     `json:"slave/cpus_revocable_used"`
-	Slave_cpusTotal                            int     `json:"slave/cpus_total"`
+	Slave_cpusRevocableTotal                   float64 `json:"slave/cpus_revocable_total"`
+	Slave_cpusRevocableUsed                    float64 `json:"slave/cpus_revocable_used"`
+	Slave_cpusTotal                            float64 `json:"slave/cpus_total"`
 	Slave_cpusUsed                             float64 `json:"slave/cpus_used"`
 	Slave_diskPercent                          float64 `json:"slave/disk_percent"`
 	Slave_diskRevocablePercent                 float64 `json:"slave/disk_revocable_percent"`
-	Slave_diskRevocableTotal                   int     `json:"slave/disk_revocable_total"`
-	Slave_diskRevocableUsed                    int     `json:"slave/disk_revocable_used"`
-	Slave_diskTotal                            int     `json:"slave/disk_total"`
-	Slave_diskUsed                             int     `json:"slave/disk_used"`
+	Slave_diskRevocableTotal                   float64 `json:"slave/disk_revocable_total"`
+	Slave_diskRevocableUsed                    float64 `json:"slave/disk_revocable_used"`
+	Slave_diskTotal                            float64 `json:"slave/disk_total"`
+	Slave_diskUsed                             float64 `json:"slave/disk_used"`
 	Slave_executorDirectoryMaxAllowedAgeSecs   float64 `json:"slave/executor_directory_max_allowed_age_secs"`
-	Slave_executorsRegistering                 int     `json:"slave/executors_registering"`
-	Slave_executorsRunning                     int     `json:"slave/executors_running"`
-	Slave_executorsTerminated                  int     `json:"slave/executors_terminated"`
-	Slave_executorsTerminating                 int     `json:"slave/executors_terminating"`
-	Slave_frameworksActive                     int     `json:"slave/frameworks_active"`
-	Slave_invalidFrameworkMessages             int     `json:"slave/invalid_framework_messages"`
-	Slave_invalidStatusUpdates                 int     `json:"slave/invalid_status_updates"`
+	Slave_executorsRegistering                 float64 `json:"slave/executors_registering"`
+	Slave_executorsRunning                     float64 `json:"slave/executors_running"`
+	Slave_executorsTerminated                  float64 `json:"slave/executors_terminated"`
+	Slave_executorsTerminating                 float64 `json:"slave/executors_terminating"`
+	Slave_frameworksActive                     float64 `json:"slave/frameworks_active"`
+	Slave_invalidFrameworkMessages             float64 `json:"slave/invalid_framework_messages"`
+	Slave_invalidStatusUpdates                 float64 `json:"slave/invalid_status_updates"`
 	Slave_memPercent                           float64 `json:"slave/mem_percent"`
 	Slave_memRevocablePercent                  float64 `json:"slave/mem_revocable_percent"`
-	Slave_memRevocableTotal                    int     `json:"slave/mem_revocable_total"`
-	Slave_memRevocableUsed                     int     `json:"slave/mem_revocable_used"`
-	Slave_memTotal                             int     `json:"slave/mem_total"`
-	Slave_memUsed                              int     `json:"slave/mem_used"`
-	Slave_recoveryErrors                       int     `json:"slave/recovery_errors"`
-	Slave_registered                           int     `json:"slave/registered"`
-	Slave_tasksFailed                          int     `json:"slave/tasks_failed"`
-	Slave_tasksFinished                        int     `json:"slave/tasks_finished"`
-	Slave_tasksKilled                          int     `json:"slave/tasks_killed"`
-	Slave_tasksLost                            int     `json:"slave/tasks_lost"`
-	Slave_tasksRunning                         int     `json:"slave/tasks_running"`
-	Slave_tasksStaging                         int     `json:"slave/tasks_staging"`
-	Slave_tasksStarting                        int     `json:"slave/tasks_starting"`
+	Slave_memRevocableTotal                    float64 `json:"slave/mem_revocable_total"`
+	Slave_memRevocableUsed                     float64 `json:"slave/mem_revocable_used"`
+	Slave_memTotal                             float64 `json:"slave/mem_total"`
+	Slave_memUsed                              float64 `json:"slave/mem_used"`
+	Slave_recoveryErrors                       float64 `json:"slave/recovery_errors"`
+	Slave_registered                           float64 `json:"slave/registered"`
+	Slave_tasksFailed                          float64 `json:"slave/tasks_failed"`
+	Slave_tasksFinished                        float64 `json:"slave/tasks_finished"`
+	Slave_tasksKilled                          float64 `json:"slave/tasks_killed"`
+	Slave_tasksLost                            float64 `json:"slave/tasks_lost"`
+	Slave_tasksRunning                         float64 `json:"slave/tasks_running"`
+	Slave_tasksStaging                         float64 `json:"slave/tasks_staging"`
+	Slave_tasksStarting                        float64 `json:"slave/tasks_starting"`
 	Slave_uptimeSecs                           float64 `json:"slave/uptime_secs"`
-	Slave_validFrameworkMessages               int     `json:"slave/valid_framework_messages"`
-	Slave_validStatusUpdates                   int     `json:"slave/valid_status_updates"`
-	System_cpusTotal                           int     `json:"system/cpus_total"`
+	Slave_validFrameworkMessages               float64 `json:"slave/valid_framework_messages"`
+	Slave_validStatusUpdates                   float64 `json:"slave/valid_status_updates"`
+	System_cpusTotal                           float64 `json:"system/cpus_total"`
 	System_load15min                           float64 `json:"system/load_15min"`
 	System_load1min                            float64 `json:"system/load_1min"`
 	System_load5min                            float64 `json:"system/load_5min"`
-	System_memFreeBytes                        int     `json:"system/mem_free_bytes"`
-	System_memTotalBytes                       int     `json:"system/mem_total_bytes"`
+	System_memFreeBytes                        float64 `json:"system/mem_free_bytes"`
+	System_memTotalBytes                       float64 `json:"system/mem_total_bytes"`
 	Time                                       time.Time
 	Node                                       string
 }
